Adding function to find particle indices

+++ AccretionTools.py
@@ -447,9 +447,9 @@ def get_particle_indices(base_halo_data,sorted_IDs,sorted_indices,IDs,Types,snap
     historyindices_atsnap=historyindices_atsnap.astype(int)
 
     #use the parttypes and history indices to find the particle data indices
-    partindices_atsnap=np.array([sorted_indices[f'PartType{ipart_type}'][ipart_historyindex] for ipart_type,ipart_historyindex in zip(parttypes_atsnap,historyindices_atsnap)])
+    partindices_atsnap=np.array([sorted_indices[f'PartType{ipart_type}'][ipart_historyindex] for ipart_type,ipart_historyindex in zip(parttypes_atsnap,historyindices_atsnap)],dtype=int)
 
-    return partindices_atsnap.astype(int),parttypes_atsnap.astype(int)
+    return partindices_atsnap,parttypes_atsnap
 
 
 ########################### GENERATE ACCRETION DATA ###########################
